refactor(game): 패키지 game 으로 분리
