+++ src/main/java/chess/domain/game/BoardFactory.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.game;
 
 import chess.domain.piece.*;
 import chess.domain.position.Column;
