+++ src/test/java/chess/domain/game/PlayerTest.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.game;
 
 import chess.domain.piece.Color;
 import chess.domain.piece.Pawn;
